requirements.txt
PyYAML
feedparser
boto3
python-dotenv
requests
